feat: CRUD usuarios, correcciones leche, config Docker/Nginx para deploy
Backend: - Agregar UsuarioController, Service, Requests (Store/Update) - Mejorar RegistroLecheController y CorregirLecheRequest - Actualizar rutas API y repositorios - Agregar Dockerfile.dev, nginx.conf y start.sh para deploy Frontend: - Mejorar UI login, home, shell y registro de leche - Agregar leche-page.scss y componente logo compartido - Actualizar leche.service.ts y estilos globales

+++ backend/app/Http/Requests/CorregirLecheRequest.php
@@ -14,7 +14,7 @@ class CorregirLecheRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'litros' => ['required', 'numeric', 'min:0'],
+            'litros' => ['required', 'integer', 'min:0'],
             'observacion' => ['required', 'string'],
         ];
     }
